Add files via upload
The Udemy dashboard created and analyzed using Power BI provides a comprehensive overview of the online courses published on the platform. The analysis revealed that IT-related courses are the most popular and highly rated among subscribers. Additionally, the average rating of the courses was higher than 4, indicating high levels of satisfaction among the subscribers. The yearly breakdown created using the slicer function helped to identify trends and patterns over time, and the drill-down function on the trend analysis of subscribers provided a detailed view of the data. Overall, the analysis provides valuable insights into the online learning market and highlights the importance of offering high-quality IT-related courses to meet the growing demand for these types of courses